docs: updated design
