+++ docs/vercel-v0-ux/LoginPage.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
-import { Checkbox } from "@/components/ui/checkbox"
+import { Card, CardContent, CardFooter, CardHeader } from "@/components/ui/card"
 
 export default function LoginPage() {
   const handleSubmit = (e: React.FormEvent) => {
@@ -12,94 +12,57 @@ export default function LoginPage() {
   }
 
   return (
-    <div className="flex items-center justify-center min-h-screen bg-background p-4">
-      <div className="w-full max-w-md space-y-8">
-        <div className="space-y-2 text-center">
+    <div className="flex items-center justify-center min-h-screen bg-background">
+      <div className="w-full max-w-md">
+        <div className="space-y-1 flex flex-col items-center mb-6">
           <Image
             src="https://hebbkx1anhila5yf.public.blob.vercel-storage.com/image-HCJ3orxGLoM9gtLGIqIDQ8CCYj6e1T.png"
             alt="IkiGoo Logo"
-            width={150}
-            height={75}
-            className="mx-auto"
+            width={200}
+            height={100}
+            className="mb-4"
           />
-          <h2 className="text-2xl font-semibold">Welcome back!</h2>
         </div>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <Input
-              id="email"
-              type="email"
-              placeholder="name@email.com"
-              required
-            />
-          </div>
-          <div className="space-y-2">
-            <Input
-              id="password"
-              type="password"
-              placeholder="Password"
-              required
-            />
-          </div>
-          <div className="flex items-center justify-between">
-            <div className="flex items-center space-x-2">
-              <Checkbox id="remember" />
-              <label htmlFor="remember" className="text-sm text-muted-foreground">
-                Remember me
-              </label>
+        <div className="space-y-6">
+          <form onSubmit={handleSubmit}>
+            <div className="space-y-4">
+              <div className="space-y-2">
+                <label htmlFor="email" className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70">
+                  Email
+                </label>
+                <Input
+                  id="email"
+                  type="email"
+                  placeholder="m@example.com"
+                  required
+                />
+              </div>
+              <div className="space-y-2">
+                <label htmlFor="password" className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70">
+                  Password
+                </label>
+                <Input
+                  id="password"
+                  type="password"
+                  required
+                />
+              </div>
+              <Button type="submit" className="w-full">
+                Sign In
+              </Button>
             </div>
-            <Link href="/reset-password" className="text-sm text-primary hover:underline">
-              Reset password
+          </form>
+          <div className="flex flex-col space-y-2 text-center">
+            <Link href="/forgot-password" className="text-sm text-muted-foreground underline hover:text-primary">
+              Forgot password?
             </Link>
+            <div className="text-sm text-muted-foreground">
+              Don't have an account?{' '}
+              <Link href="/signup" className="underline hover:text-primary">
+                Sign up
+              </Link>
+            </div>
           </div>
-          <Button type="submit" className="w-full bg-black text-white hover:bg-gray-800">
-            Sign in
-          </Button>
-        </form>
-        <div className="text-center text-sm">
-          Don't have an account?{' '}
-          <Link href="/signup" className="text-primary font-semibold hover:underline">
-            Sign up
-          </Link>
-        </div>
-        <div className="relative">
-          <div className="absolute inset-0 flex items-center">
-            <span className="w-full border-t" />
-          </div>
-          <div className="relative flex justify-center text-xs uppercase">
-            <span className="bg-background px-2 text-muted-foreground">
-              Or continue with
-            </span>
-          </div>
-        </div>
-        <div className="space-y-2">
-          <Button variant="outline" className="w-full">
-            <svg className="w-5 h-5 mr-2" viewBox="0 0 24 24">
-              <path
-                fill="currentColor"
-                d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-              />
-              <path
-                fill="currentColor"
-                d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-              />
-              <path
-                fill="currentColor"
-                d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-              />
-              <path
-                fill="currentColor"
-                d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-              />
-            </svg>
-            Continue with Google
-          </Button>
-          <Button variant="outline" className="w-full">
-            <svg className="w-5 h-5 mr-2 text-[#1DA1F2]" fill="currentColor" viewBox="0 0 24 24">
-              <path d="M22.46 6c-.77.35-1.6.58-2.46.69.88-.53 1.56-1.37 1.88-2.38-.83.5-1.75.85-2.72 1.05C18.37 4.5 17.26 4 16 4c-2.35 0-4.27 1.92-4.27 4.29 0 .34.04.67.11.98C8.28 9.09 5.11 7.38 3 4.79c-.37.63-.58 1.37-.58 2.15 0 1.49.75 2.81 1.91 3.56-.71 0-1.37-.2-1.95-.5v.03c0 2.08 1.48 3.82 3.44 4.21a4.22 4.22 0 0 1-1.93.07 4.28 4.28 0 0 0 4 2.98 8.521 8.521 0 0 1-5.33 1.84c-.34 0-.68-.02-1.02-.06C3.44 20.29 5.7 21 8.12 21 16 21 20.33 14.46 20.33 8.79c0-.19 0-.37-.01-.56.84-.6 1.56-1.36 2.14-2.23z" />
-            </svg>
-            Continue with Twitter
-          </Button>
         </div>
       </div>
     </div>
